Add files via upload
This assignment was quite a challenge! I ended up talking to my tutor to figure out why my code wouldn't loop through all the changes- it turns out it was a scope issue, and I just needed to set a worksheet variable and deploy it on my Cells, Range, and Columns properties to get the code to iterate through all worksheets, even though the rest of the For Each loop to do so was already in place. In all honesty, I don't understand why the worksheets variable has to be attached to these properties, but I do understand that it needs to be there. In my research, I stumbled over the concept of Offset and CurrentRegion properties, which I wish I'd known before I started writing the code as it would have prevented some of my simple formatting errors early on in the project.

Perhaps the biggest idiosyncrasy of this project is found in lines 53-55. I know there's a way to retrieve the value I was seeking by using the If Cells(x,y) <> Cells(x,y+1) sort of thing, but for the life of me i couldn't get it to work on the first line of code in the project. When I tried to use my Google skills to find a solution, I stumbled over a few sites where former students of this class had uploaded the entire assignment to a help site and asked for input. I stayed clear of these pages, but had a hard time finding a solution anywhere else, so I rubbed all six of my brain cells together until I came up with the concept of the "nope"/"yup" checker. I'm certain I'm neither the first nor last person to use this method, so I make no claims of discovery, but I'm proud of myself for coming to that solution on my own.

Another topic I discussed with my tutor was how to use GitHub at all, as this was my first experience doing so. Since I didn't talk to him until my homework was basically finished, I only ended up doing two commits for this project. In the future, I'll try to do more commits on my projects.